Create utility function for account controller
