+++ controllers/account-controller.js
@@ -10,6 +10,8 @@ const bcrypt = require('bcrypt');
 /* Use ten salt rounds for password hashing. */
 const saltRounds = 10;
 
+const accountControllerUtil = require('./account-controller-util.js');
+
 const accountController = {
 	/**
 	 * Gets the account page.
@@ -25,25 +27,7 @@ const accountController = {
 			const projection = '_id firstName lastName username role status';
 
 			db.findMany(Account, query, projection, function(result) {
-				/* Assign the result of the database retrieval to the variable accounts. */
-				const accounts = result;
-
-				/* For each account, push its details to the accountDetails array if the account
-				 * does not belong to the administrator.
-				 */
-				const accountDetails = [];
-				for (let i = 0; i < accounts.length; i++) {
-					if (accounts[i].role != 'administrator') {
-						accountDetails.push(accounts[i]);
-					}
-				}
-
-				/* Store the retrieved account details in the variable data. */
-				const data = {
-					accountDetails: accountDetails
-				};
-
-				res.render('account', data);
+				res.render('account', accountControllerUtil.accountUtil(result));
 			});
 		} else {
 			/* Retrieve the account details of the logged in user if a regular user accesses the account tab. */
